optimize field in database by using index and smaller type

+++ database/migrations/2025_12_20_073144_create_comment_attachments_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('path');
             $table->string('mime_type')->nullable();
-            $table->unsignedBigInteger('size')->default(0);
+            $table->unsignedInteger('size')->default(0);
             $table->timestamps();
         });
     }
